selectAllCustomers, selectCustomerById tests. Change from Integer to Long (for ID)

+++ src/main/java/com/amigoscode/customer/CustomerController.java
@@ -21,7 +21,7 @@ public class CustomerController {
 
     @GetMapping("{customerId}")
     public Customer getCustomer(
-            @PathVariable("customerId") Integer customerId) {
+            @PathVariable("customerId") Long customerId) {
         return customerService.getCustomer(customerId);
     }
 
@@ -33,13 +33,13 @@ public class CustomerController {
 
     @DeleteMapping("{customerId}")
     public void deleteCustomer(
-            @PathVariable("customerId") Integer customerId) {
+            @PathVariable("customerId") Long customerId) {
         customerService.deleteCustomerById(customerId);
     }
 
     @PutMapping("{customerId}")
     public void updateCustomer(
-            @PathVariable("customerId") Integer customerId,
+            @PathVariable("customerId") Long customerId,
             @RequestBody CustomerUpdateRequest updateRequest) {
         customerService.updateCustomer(customerId, updateRequest);
     }
