+++ src/main/java/com/amigoscode/customer/CustomerJDBCDataAccessService.java
@@ -27,7 +27,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public Optional<Customer> selectCustomerById(Integer id) {
+    public Optional<Customer> selectCustomerById(Long id) {
         var sql = """
                 SELECT id, name, email, age
                 from customer
@@ -64,7 +64,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public boolean existsPersonWithId(Integer id) {
+    public boolean existsPersonWithId(Long id) {
         var sql = """
                 SELECT count(id)
                 from customer
@@ -75,7 +75,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public void deleteCustomerById(Integer id) {
+    public void deleteCustomerById(Long id) {
         var sql = """
                 DELETE FROM customer
                 WHERE id = ?
