+++ src/main/java/com/amigoscode/customer/CustomerService.java
@@ -21,7 +21,7 @@ public class CustomerService {
         return customerDao.selectAllCustomers();
     }
 
-    public Customer getCustomer(Integer id) {
+    public Customer getCustomer(Long id) {
         return customerDao.selectCustomerById(id)
                 .orElseThrow(() -> new ResourceNotFoundException(
                         "customer with id [%s] not found".formatted(id)
@@ -44,7 +44,7 @@ public class CustomerService {
         customerDao.insertCustomer(customer);
     }
 
-    public void deleteCustomerById(Integer id) {
+    public void deleteCustomerById(Long id) {
         if (!customerDao.existsPersonWithId(id)) {
             throw new ResourceNotFoundException(
                     "customer with id [%s] not found".formatted(id)
@@ -54,7 +54,7 @@ public class CustomerService {
     }
 
     public void updateCustomer(
-            Integer customerId,
+            Long customerId,
             CustomerUpdateRequest updateRequest) {
         var customer = getCustomer(customerId);
